combined algorithm and lib
